♻️ [Refactoring]: #176 PdfType の直接 import を新パスへ切り替え
catalog-parser.ts と object-stream-extractor/dict/index.ts の 2 ファイルから
旧 pdf/type/index への直接 import を新 pdf/types/pdf-type/index へ更新する。
実装ロジックは未変更。
